Add mapping from data models to domain models

+++ app/src/main/java/ru/zarina/zarina/domain/rework/product/Price.kt
@@ -1,10 +1,8 @@
 package ru.zarina.zarina.domain.rework.product
 
-import android.icu.math.BigDecimal
-
 data class Price(
-    val commonPrice: BigDecimal,
+    val commonPrice: Long,
     val hasDiscount: Boolean,
-    val discountPrice: BigDecimal,
+    val discountPrice: Long,
     val discountPercent: Int,
 )
